+++ secondhand/src/test/java/com/codesquad/secondhand/domain/member/service/MemberServiceTest.java
@@ -1,6 +1,5 @@
 package com.codesquad.secondhand.domain.member.service;
 
-
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.List;
@@ -12,11 +11,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.redis.core.RedisTemplate;
 
 import com.codesquad.secondhand.annotation.ServiceIntegrationTest;
+import com.codesquad.secondhand.domain.jwt.repository.TokenJpaRepository;
 import com.codesquad.secondhand.domain.member.dto.request.SignupRequest;
 import com.codesquad.secondhand.domain.member.entity.Member;
 import com.codesquad.secondhand.domain.member_region.entity.MemberRegion;
 import com.codesquad.secondhand.domain.member_region.service.MemberRegionQueryService;
-import com.codesquad.secondhand.domain.token.repository.TokenJpaRepository;
 import com.codesquad.secondhand.exception.CustomRuntimeException;
 
 @ServiceIntegrationTest
@@ -37,22 +36,21 @@ class MemberServiceTest {
 	@Autowired
 	TokenJpaRepository tokenJpaRepository;
 
-
-
 	@AfterEach
 	void resetBlackList() {
-		redisBlackListTemplate.delete("eyJhbGciOiJIUzI1NiJ9.eyJlbWFpbCI6InRlc3RAdGVzdC5jb20iLCJtZW1iZXJJZCI6MSwiY3JlYXRlZE1pbGxpcyI6MTY5Mzc5NjE4NTE0NiwiZXhwIjoxNzI1MzMyMTg1fQ.QKKVe4mEOOO8hzVsQ5xyDghyQluxhOAWyNfV3vIJSfI");
+		redisBlackListTemplate.delete(
+			"eyJhbGciOiJIUzI1NiJ9.eyJlbWFpbCI6InRlc3RAdGVzdC5jb20iLCJtZW1iZXJJZCI6MSwiY3JlYXRlZE1pbGxpcyI6MTY5Mzc5NjE4NTE0NiwiZXhwIjoxNzI1MzMyMTg1fQ.QKKVe4mEOOO8hzVsQ5xyDghyQluxhOAWyNfV3vIJSfI");
 	}
 
 	@DisplayName("회원가입에 성공한다.")
 	@Test
 	void signUpSuccess() {
 		// given
-		List<Long> ids = List.of(1L,2L);
-		SignupRequest signupRequest = new SignupRequest("test","test", ids);
+		List<Long> ids = List.of(1L, 2L);
+		SignupRequest signupRequest = new SignupRequest("test", "test", ids);
 		String requestEmail = "test@email.com";
 		// when
-		memberService.signUp(signupRequest,requestEmail);
+		memberService.signUp(signupRequest, requestEmail);
 		Member savedMember = memberQueryService.findByEmail(requestEmail).get();
 		List<MemberRegion> memberRegions = memberRegionQueryService.findAllMemberRegion(savedMember.getId());
 		// then
@@ -66,14 +64,14 @@ class MemberServiceTest {
 	@Test
 	void signUpFailed() {
 		// given
-		List<Long> ids = List.of(1L,2L);
-		SignupRequest signupRequest = new SignupRequest("test","test", ids);
+		List<Long> ids = List.of(1L, 2L);
+		SignupRequest signupRequest = new SignupRequest("test", "test", ids);
 		String requestEmail = "test@email.com";
-		memberService.signUp(signupRequest,requestEmail);
+		memberService.signUp(signupRequest, requestEmail);
 		String savedMemberNickname = memberQueryService.findByEmail(requestEmail).get().getNickname();
-		SignupRequest existNicknameRequest = new SignupRequest(savedMemberNickname,"test", ids);
+		SignupRequest existNicknameRequest = new SignupRequest(savedMemberNickname, "test", ids);
 		// when & then
-		assertThatThrownBy(() -> memberService.signUp(existNicknameRequest,requestEmail)).isInstanceOf(
+		assertThatThrownBy(() -> memberService.signUp(existNicknameRequest, requestEmail)).isInstanceOf(
 			CustomRuntimeException.class);
 	}
 }
